+++ lib/pkg.ts
@@ -1,4 +1,4 @@
 export default {
-  version: '1.2.0-beta.10',
+  version: '1.2.0-beta.11',
   name: 'aliyun-pds-js-sdk',
 }